DBZ-3099 Address review comments; extract history comparator for testability
